feat: splitting view model out
