+++ Sources/adadapted-swift-sdk/core/view/AdZonePresenter.swift
@@ -46,7 +46,6 @@ class AdZonePresenter: SessionListener {
             attached = true
             self.adZonePresenterListener = adZonePresenterListener
             sessionClient?.addPresenter(listener: self)
-            //setNextAd() TODO
         }
     }
